Add verification params Id and userId via token  is the same Import bcrypt (was missing for updating password)

+++ server/controllers/userController.js
@@ -1,7 +1,9 @@
 import User from '../models/userModel.js'
+import bcrypt from 'bcrypt'
 
 export const getAllUsers = async (req, res) => {
   try {
+    // Order by most recent
     const users = await User.find({}).sort({ createdAt: -1 })
 
     return res.status(200).json(users)
@@ -12,6 +14,11 @@ export const getAllUsers = async (req, res) => {
 
 export const getUser = async (req, res) => {
   const id = req.params.id
+  const tokenUserId = req.userId
+
+  if (id !== tokenUserId) {
+    return res.status(403).json({ message: 'Not Authorized!' })
+  }
 
   try {
     const user = await User.findById({ _id: id })
@@ -28,6 +35,11 @@ export const getUser = async (req, res) => {
 
 export const updateUser = async (req, res) => {
   const id = req.params.id
+  const tokenUserId = req.userId
+
+  if (id !== tokenUserId) {
+    return res.status(403).json({ message: 'Not Authorized!' })
+  }
 
   try {
     if (req.body.password) {
@@ -48,13 +60,18 @@ export const updateUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
   const id = req.params.id
+  const tokenUserId = req.userId
+
+  if (id !== tokenUserId) {
+    return res.status(403).json({ message: 'Not Authorized!' })
+  }
 
   try {
     const user = await User.findOneAndDelete({ _id: id })
     if (!user) {
       return res.status(400).json({ error: 'No such user' })
     }
-    res.status(200).json(user)
+    res.status(200).json({ message: 'User deleted successfuly' })
   } catch (error) {
     return res.status(500).json({ error: error.message })
   }
